Refactor del mapa de sucursales
